converted random number return type to double

+++ Monte_Carlo.cpp
@@ -1,7 +1,6 @@
 #include "headers.h"
 
-Monte_Carlo :: Monte_Carlo(function<vector<int> (int, int, pair<int, int>)> generate_random_numbers)
-// Monte_Carlo :: Monte_Carlo(vector<int> generate_random_numbers(int, int, pair<int, int>))
+Monte_Carlo :: Monte_Carlo(function<vector<double> (int, int, pair<int, int>)> generate_random_numbers)
 {
     random_num_gen = generate_random_numbers;
 }
@@ -9,6 +8,6 @@ Monte_Carlo :: Monte_Carlo(function<vector<int> (int, int, pair<int, int>)> gene
 double Monte_Carlo :: generate_pi()
 {
     pair<int, int> r = make_pair(6,1);
-    vector<int> tmp = random_num_gen(5, 7, r);
+    vector<double> tmp = random_num_gen(5, 7, r);
     return 3.14;
 }
