+++ lcg_parallel1.h
@@ -8,7 +8,7 @@ class LCG_Parallel1 : public LCG
 public:
 	int num_procs;
 	LCG_Parallel1(int number_of_processes);
-	vector<int> generate_random_numbers(int count, int seed, pair<int, int> range);
+	vector<double> generate_random_numbers(int count, int seed, pair<int, int> range);
 };
 
 #endif
